+++ src/component/mod.rs
@@ -17,105 +17,26 @@ pub(crate) use seal::Component as SealedComponent;
 ///
 /// Note that not all components are necessarily transformable; most colour spaces and formats require some form of floating- or fixed-point arithmetic for transformations to be at least be somewhat accurate.
 /// Only the most basic of colours spaces can thus be translated using simple integers.
-pub trait Component: SealedComponent + Copy + Sized {
-	/// Lossily converts the component to [`f16`].
-	#[cfg(feature = "f16")]
-	#[must_use]
-	fn to_f16_lossy(self) -> f16;
+pub trait Component: SealedComponent + Copy + Sized { }
 
-	/// Lossily converts the component to [`f32`].
-	#[must_use]
-	fn to_f32_lossy(self) -> f32;
-
-	/// Lossily converts the component to [`f64`].
-	#[must_use]
-	fn to_f64_lossy(self) -> f64;
-
-	/// Lossily converts the component to [`f128`].
-	#[cfg(feature = "f128")]
-	#[must_use]
-	fn to_f128_lossy(self) -> f128;
-}
-
-macro_rules! impl_integer_component {
+macro_rules! impl_component {
 	{
 		$(
 			$(#[$attrs:meta])*
 			$tys:ty
-		),+$(,)?
+		),*$(,)?
 	} => {
 		$(
 			$(#[$attrs])*
 			impl ::polywave::SealedComponent for $tys { }
 
 			$(#[$attrs])*
-			impl ::polywave::Component for $tys {
-				#[cfg(feature = "f16")]
-				#[inline(always)]
-				fn to_f16_lossy(self) -> f16 {
-					self as f16 / Self::MAX as f16
-				}
-
-				#[inline(always)]
-				fn to_f32_lossy(self) -> f32 {
-					self as f32 / Self::MAX as f32
-				}
-
-				#[inline(always)]
-				fn to_f64_lossy(self) -> f64 {
-					self as f64 / Self::MAX as f64
-				}
-
-				#[cfg(feature = "f128")]
-				#[inline(always)]
-				fn to_f128_lossy(self) -> f128 {
-					self as f128 / Self::MAX as f128
-				}
-			}
+			impl ::polywave::Component for $tys { }
 		)*
 	};
 }
 
-macro_rules! impl_float_component {
-	{
-		$(
-			$(#[$attrs:meta])*
-			$tys:ty
-		),+$(,)?
-	} => {
-		$(
-			$(#[$attrs])*
-			impl ::polywave::SealedComponent for $tys { }
-
-			$(#[$attrs])*
-			impl ::polywave::Component for $tys {
-				#[cfg(feature = "f16")]
-				#[inline(always)]
-				fn to_f16_lossy(self) -> f16 {
-					self as f16
-				}
-
-				#[inline(always)]
-				fn to_f32_lossy(self) -> f32 {
-					self as f32
-				}
-
-				#[inline(always)]
-				fn to_f64_lossy(self) -> f64 {
-					self as f64
-				}
-
-				#[cfg(feature = "f128")]
-				#[inline(always)]
-				fn to_f128_lossy(self) -> f128 {
-					self as f128
-				}
-			}
-		)*
-	};
-}
-
-impl_integer_component! {
+impl_component! {
 	u8,
 	i8,
 	u16,
@@ -126,9 +47,7 @@ impl_integer_component! {
 	i64,
 	u128,
 	i128,
-}
 
-impl_float_component! {
 	f32,
 	f64,