+++ src/hsv/hsva/mod.rs
@@ -31,9 +31,11 @@ impl<T: Component> Hsva<T> {
 	/// Discards the HSVA colour's alpha channel.
 	#[inline(always)]
 	#[must_use]
-	pub const fn discard_alpha(self) -> Hsv<T> {
-		let (hue, saturation, value, _) = self.get();
-		Hsv::new(hue, saturation, value)
+	pub const fn discard_alpha(self) -> (Hsv<T>, T) {
+		let (hue, saturation, value, alpha) = self.get();
+
+		let colour = Hsv::new(hue, saturation, value);
+		(colour, alpha)
 	}
 
 	/// Deconstructs an HSVA colour.
@@ -44,5 +46,39 @@ impl<T: Component> Hsva<T> {
 		(hue, saturation, value, alpha)
 	}
 }
+
+macro_rules! impl_conversions {
+	($($tys:ty),+$(,)?) => {
+		$(
+			impl ::polywave::hsv::Hsva<$tys> {
+				/// Converts an HWBA colour to HSVA.
+				pub const fn from_hwba(colour: ::polywave::hsv::Hwba<$tys>) -> Self {
+					let (colour, alpha) = colour.discard_alpha();
+
+					let colour = Hsv::<$tys>::from_hwb(colour);
+
+					colour.with_alpha(alpha)
+				}
+
+				/// Converts the HSVA colour to HWBA.
+				pub fn to_hwba(self) -> ::polywave::hsv::Hwba<$tys> {
+					let (colour, alpha) = self.discard_alpha();
+
+					let colour = colour.to_hwb();
+
+					colour.with_alpha(alpha)
+				}
+			}
+		)*
+	};
+}
+
+impl_conversions!(f32, f64);
+
+#[cfg(feature = "f16")]
+impl_conversions!(f16);
+
+#[cfg(feature = "f128")]
+impl_conversions!(f128);
 
 impl<T: Component> Colour for Hsva<T> { }