+++ src/hsv/hwba/mod.rs
@@ -1,0 +1,50 @@
+// Copyright 2025 Gabriel Bjørnager Jensen.
+//
+// This Source Code Form is subject to the terms of
+// the Mozilla Public License, v. 2.0. If a copy of
+// the MPL was not distributed with this file, you
+// can obtain one at:
+// <https://mozilla.org/MPL/2.0/>.
+
+use crate::{Colour, Component};
+use crate::hsv::Hwb;
+
+/// An HWBA colour.
+///
+/// This type guarantees that its four channels -- hue, whiteness, blackness, and alpha -- are stored sequentially in memory (in this order).
+#[repr(transparent)]
+#[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
+#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
+#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+pub struct Hwba<T>([T; 0x4]);
+
+impl<T: Component> Hwba<T> {
+	/// Constructs a new HWBA colour.
+	#[inline(always)]
+	#[must_use]
+	pub const fn new(hue: T, whiteness: T, blackness: T, alpha: T) -> Self {
+		let data = [hue, whiteness, blackness, alpha];
+		Self(data)
+	}
+
+	/// Discards the HWBA colour's alpha channel.
+	#[inline(always)]
+	#[must_use]
+	pub const fn discard_alpha(self) -> (Hwb<T>, T) {
+		let (hue, whiteness, blackness, alpha) = self.get();
+
+		let colour = Hwb::new(hue, whiteness, blackness);
+		(colour, alpha)
+	}
+
+	/// Deconstructs an HWBA colour.
+	#[inline(always)]
+	#[must_use]
+	pub const fn get(self) -> (T, T, T, T) {
+		let [hue, whiteness, blackness, alpha] = self.0;
+		(hue, whiteness, blackness, alpha)
+	}
+}
+
+impl<T: Component> Colour for Hwba<T> { }