+++ src/hsv/mod.rs
@@ -8,8 +8,16 @@
 
 //! HSV- and HSL-based colour facilities.
 
+mod hsl;
+mod hsla;
 mod hsv;
 mod hsva;
+mod hwb;
+mod hwba;
 
+pub use hsl::Hsl;
+pub use hsla::Hsla;
 pub use hsv::Hsv;
 pub use hsva::Hsva;
+pub use hwb::Hwb;
+pub use hwba::Hwba;