+++ src/lch/mod.rs
@@ -6,14 +6,14 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-//! LAB-based colour facilities.
+//! LCh-based colour facilities.
 
-mod cie_lab;
-mod cie_laba;
-mod ok_lab;
-mod ok_laba;
+mod cie_lch;
+mod cie_lcha;
+mod ok_lch;
+mod ok_lcha;
 
-pub use cie_lab::CieLab;
-pub use cie_laba::CieLaba;
-pub use ok_lab::OkLab;
-pub use ok_laba::OkLaba;
+pub use cie_lch::CieLch;
+pub use cie_lcha::CieLcha;
+pub use ok_lch::OkLch;
+pub use ok_lcha::OkLcha;