+++ src/lch/ok_lcha/mod.rs
@@ -7,46 +7,48 @@
 // <https://mozilla.org/MPL/2.0/>.
 
 use crate::{Colour, Component, DefinedGamut};
-use crate::lab::CieLch;
+use crate::lch::OkLch;
 
-/// A CIELChA colour.
+/// An Oklcha colour.
 ///
 /// This type guarantees that its three channels -- luminance, chroma, hue, and alpha -- are stored sequentially in memory (in this order).
 ///
-/// This type is equivalent to [`CieLch`] with an added alpha component.
+/// This type is equivalent to [`OkLch`] with an added alpha component.
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
 #[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
 #[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
-pub struct CieLcha<T>([T; 0x4]);
+pub struct OkLcha<T>([T; 0x4]);
 
-impl<T: Component> CieLcha<T> {
-	/// Constructs a new CIELChA colour.
+impl<T: Component> OkLcha<T> {
+	/// Constructs a new Oklcha colour.
 	#[inline(always)]
 	#[must_use]
-	pub const fn new(x: T, y: T, z: T, alpha: T) -> Self {
-		let data = [x, y, z, alpha];
+	pub const fn new(lightness: T, chroma: T, hue: T, alpha: T) -> Self {
+		let data = [lightness, chroma, hue, alpha];
 		Self(data)
 	}
 
-	/// Discards the CIELChA colour's alpha channel.
+	/// Discards the Oklcha colour's alpha channel.
 	#[inline(always)]
 	#[must_use]
-	pub const fn discard_alpha(self) -> CieLch<T> {
-		let (x, y, z, _) = self.get();
-		CieLch::new(x, y, z)
+	pub const fn discard_alpha(self) -> (OkLch<T>, T) {
+		let (lightness, chroma, hue, alpha) = self.get();
+
+		let colour = OkLch::new(lightness, chroma, hue);
+		(colour, alpha)
 	}
 
-	/// Deconstructs a CIELChA colour.
+	/// Deconstructs an Oklcha colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn get(self) -> (T, T, T, T) {
-		let [x, y, z, alpha] = self.0;
-		(x, y, z, alpha)
+		let [lightness, chroma, hue, alpha] = self.0;
+		(lightness, chroma, hue, alpha)
 	}
 }
 
-impl<T: Component> Colour for CieLcha<T> { }
+impl<T: Component> Colour for OkLcha<T> { }
 
-impl<T: Component> DefinedGamut for CieLcha<T> { }
+impl<T: Component> DefinedGamut for OkLcha<T> { }