+++ src/lib.rs
@@ -26,6 +26,7 @@ extern crate alloc;
 pub mod error;
 pub mod hsv;
 pub mod lab;
+pub mod lch;
 pub mod rgb;
 pub mod www;
 pub mod xyz;