+++ src/rgb/s_rgba/mod.rs
@@ -34,9 +34,11 @@ impl<T: Component> SRgba<T> {
 	/// Discards the sRGBA colour's alpha channel.
 	#[inline(always)]
 	#[must_use]
-	pub const fn discard_alpha(self) -> SRgb<T> {
-		let (red, green, blue, _) = self.get();
-		SRgb::new(red, green, blue)
+	pub const fn discard_alpha(self) -> (SRgb<T>, T) {
+		let (red, green, blue, alpha) = self.get();
+
+		let colour = SRgb::new(red, green, blue);
+		(colour, alpha)
 	}
 
 	/// Reinterprets the sRGBA colour as a raw RGB colour.